pyproject.toml
[project]
name = "simplematrixbotlib"
dynamic = ["version"]
maintainers = [
  {name = "Isaac Beverly", email = "imbev@protonmail.com"},
]
description = "A simple Python bot library for the Matrix Protocol"
readme = "README.md"
license = {file = "LICENSE"}
keywords = [
    "matrix",
    "bot",
    "library",
    "simplematrixbotlib",
    "simple-matrix-bot-lib",
]
classifiers = [
    "Topic :: Communications :: Chat :: Matrix",
    "Topic :: Internet",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Intended Audience :: Developers",
]
dependencies = [
  "matrix-nio",
  "pillow",
  "markdown",
]

[project.optional-dependencies]
dev = [
    "build",
    "twine"
]
docs = [
    "pdoc"
]

[project.urls]
Documentation = "https://simplematrixbotlib.readthedocs.io"
Repository = "https://codeberg.org/imbev/simplematrixbotlib"
Issues = "https://codeberg.org/imbev/simplematrixbotlib/issues"
Chat = "https://matrix.to/#/#simplematrixbotlib:matrix.org"

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["simplematrixbotlib"]

[tool.setuptools.dynamic]
version = {attr = "simplematrixbotlib.__version__"}
